requirements.txt
# GUI依赖
PySide6>=6.4.0
mysql-connector-python>=8.0.0

# Web框架依赖
fastapi>=0.111,<0.116
uvicorn[standard]>=0.30,<0.33
sqlalchemy>=2.0.30,<2.1
pymysql>=1.1,<2
pydantic>=2.7,<3
httpx>=0.27,<0.28

# 爬虫相关依赖
trafilatura>=1.8,<2
beautifulsoup4>=4.12,<5
requests>=2.28.0
lxml>=4.9.0

# 其他依赖
jinja2>=3.1,<4
python-dotenv>=1.0,<2